refactor: move X11 global hotkey out of hotkeywrapper.cc
